requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# LangChain & LLM
langchain>=0.3.7
langchain-community>=0.3.7
langchain-core>=0.3.19
langchain-openai>=0.2.8
langchain-ollama>=0.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# HTTP Client
httpx>=0.25.0

# Configuration
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0
